7 Added Data and constructor for Model and changed for Entity.
