+++ DB-Connector/src/main/java/com/pharos/dbconnector/entity/Document.java
@@ -1,16 +1,11 @@
-package com.pharos.dbconnector.model;
+package com.pharos.dbconnector.entity;
 
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Column;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.NoArgsConstructor;
-import lombok.Builder;
+import lombok.*;
 
 @Entity
 @AllArgsConstructor
@@ -18,14 +13,15 @@ import lombok.Builder;
 @Setter
 @ToString
 @NoArgsConstructor
+@Data
 @Builder
-public class Gender {
+public class Document {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "GENDER_ID", nullable = false)
-    private int idGender;
-    private String genderVal;
+    @Column(name = "ID_DOC", nullable = false)
+    private int idDoc;
+    private String documentId;
     private Boolean enable;
 
 }
